// src/types.ts
export type IssueStatus = 'Backlog' | 'In Progress' | 'Done';
export type IssuePriority = 'low' | 'medium' | 'high';

export interface Issue {
  id: string;
  title: string;
  status: IssueStatus;
  priority: IssuePriority;
  createdAt: string;
  severity: number;
  assignee: string;
  tags: string[];
  userDefinedRank?: number;
}